Include Git revision as version
Updated Makefile and main.c to include a git version string if the program is
compiled within a git repository. The version string will be formatted like
"git:<rev>" where <rev> is the output of `git describe --always --tags`. If
`git describe` doesn't return anything (as is the case when we're not in a Git
repo), the version string is empty and will not be included in the compilation.

If the version string is passed to the program by a defined called GITVERSiON
(must be a quoted string), it will be compiled and printed on the help page
(when the program is called like `./stablefinder -h`). That define can be
entered on the compiler's command line (as happens in the Makefile) or defined
in some external file that gets included in main.c.