Fix sha1 call to actually do something
Since a very early revision, the call to sha1_buffer() used a buffer length
argument of '0' (where obviously it should have been 20), because of which the
function would always return the same hash. As we have not previously looked at
the actual return values of the function, we hadn't noticed this before.